Add win loss point plot into rows of a gt table

The gt_plt_winloss function takes an existing gt_tbl object and adds squares of a specific color and vertical position based on wins/losses. It is a wrapper around gt::text_transform().

gt_plt_winloss(
  gt_object,
  column,
  max_wins = 17,
  colors = c("#D50A0A", "#013369", "gray"),
  type = "pill"
)

Arguments

gt_object
An existing gt table object of class gt_tbl

column
The column wherein the winloss plot should replace existing data. Note that the data must be represented as a list of numeric values ahead of time.

max_wins
An integer indicating the max possible wins, this will be used to add padding if the total wins/losses observed is less than the max. This is useful for mid-season reporting. Defaults to a red, blue, grey palette.

colors
A character vector of length 3, specifying the colors for loss, win, tie in that exact order.

type
A character string representing the type of plot, either a 'pill' or 'square'

Value

An object of class gt_tbl.

Function ID

1-3

See also

Examples